+++ packages/jazz-tools/src/tools/coValues/interfaces.ts
@@ -4,7 +4,7 @@ import {
   type CojsonInternalTypes,
   type RawCoValue,
 } from "cojson";
-import { AvailableCoValueCore } from "cojson/dist/coValueCore/coValueCore.js";
+import { AvailableCoValueCore } from "cojson";
 import {
   Account,
   AnonymousJazzAgent,
@@ -13,7 +13,9 @@ import {
   NotLoadedCoValueState,
   type Group,
   Loaded,
+  Inaccessible,
   MaybeLoaded,
+  Settled,
   RefsToResolve,
   RefsToResolveStrict,
   RegisteredSchemas,
@@ -21,7 +23,6 @@ import {
   ResolveQueryStrict,
   Resolved,
   SubscriptionScope,
-  type SubscriptionValue,
   TypeSym,
   NotLoaded,
   activeAccountContext,
@@ -29,7 +30,7 @@ import {
   inspect,
 } from "../internal.js";
 import type { BranchDefinition } from "../subscribe/types.js";
-import { CoValueHeader } from "cojson/dist/coValueCore/verifiedState.js";
+import { CoValueHeader } from "cojson";
 
 /** @category Abstract interfaces */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -145,7 +146,7 @@ export function loadCoValueWithoutMe<
     skipRetry?: boolean;
     unstable_branch?: BranchDefinition;
   },
-): Promise<MaybeLoaded<Resolved<V, R>>> {
+): Promise<Settled<Resolved<V, R>>> {
   return loadCoValue(cls, id, {
     ...options,
     loadAs: options?.loadAs ?? activeAccountContext.get(),
@@ -165,7 +166,7 @@ export function loadCoValue<
     skipRetry?: boolean;
     unstable_branch?: BranchDefinition;
   },
-): Promise<MaybeLoaded<Resolved<V, R>>> {
+): Promise<Settled<Resolved<V, R>>> {
   return new Promise((resolve) => {
     subscribeToCoValue<V, R>(
       cls,
@@ -302,8 +303,8 @@ export function subscribeToCoValue<
   options: {
     resolve?: RefsToResolveStrict<V, R>;
     loadAs: Account | AnonymousJazzAgent;
-    onUnavailable?: (value: NotLoaded<V>) => void;
-    onUnauthorized?: (value: NotLoaded<V>) => void;
+    onUnavailable?: (value: Inaccessible<V>) => void;
+    onUnauthorized?: (value: Inaccessible<V>) => void;
     syncResolution?: boolean;
     skipRetry?: boolean;
     unstable_branch?: BranchDefinition;
@@ -342,16 +343,10 @@ export function subscribeToCoValue<
 
     switch (value.$jazz.loadingState) {
       case CoValueLoadingState.UNAVAILABLE:
-        options.onUnavailable?.(value);
-
-        // Don't log unavailable errors when `loadUnique` or `upsertUnique` are used
-        if (!options.skipRetry) {
-          console.error(value.toString());
-        }
+        options.onUnavailable?.(value as Inaccessible<V>);
         break;
       case CoValueLoadingState.UNAUTHORIZED:
-        options.onUnauthorized?.(value);
-        console.error(value.toString());
+        options.onUnauthorized?.(value as Inaccessible<V>);
         break;
     }
   };
@@ -546,7 +541,7 @@ export async function internalLoadUnique<
     owner: Account | Group;
     resolve?: RefsToResolveStrict<V, R>;
   },
-): Promise<MaybeLoaded<Resolved<V, R>>> {
+): Promise<Settled<Resolved<V, R>>> {
   const loadAs = options.owner.$jazz.loadedAs;
 
   const node =
@@ -682,18 +677,18 @@ export async function exportCoValue<
   );
 
   const value = await new Promise<Loaded<S, R> | null>((resolve) => {
-    rootNode.setListener((value) => {
-      if (value.type === CoValueLoadingState.UNAVAILABLE) {
-        resolve(null);
-        console.error(value.toString());
-      } else if (value.type === CoValueLoadingState.UNAUTHORIZED) {
-        resolve(null);
-        console.error(value.toString());
-      } else if (value.type === CoValueLoadingState.LOADED) {
-        resolve(value.value as Loaded<S, R>);
-      }
+    rootNode.setListener(() => {
+      const value = rootNode.getCurrentValue();
 
-      rootNode.destroy();
+      if (value.$isLoaded) {
+        resolve(value as Loaded<S, R>);
+      } else if (
+        value.$jazz.loadingState === CoValueLoadingState.UNAVAILABLE ||
+        value.$jazz.loadingState === CoValueLoadingState.UNAUTHORIZED
+      ) {
+        resolve(null);
+        rootNode.destroy();
+      }
     });
   });