+++ packages/jazz-tools/src/tools/implementation/zodSchema/schemaTypes/AccountSchema.ts
@@ -5,7 +5,7 @@ import {
   CoMapSchemaDefinition,
   coOptionalDefiner,
   Group,
-  MaybeLoaded,
+  Settled,
   RefsToResolveStrict,
   RefsToResolve,
   Resolved,
@@ -26,6 +26,7 @@ import {
 import { CoOptionalSchema } from "./CoOptionalSchema.js";
 import { CoreResolveQuery } from "./CoValueSchema.js";
 import { withSchemaResolveQuery } from "../../schemaUtils.js";
+import { AgentSecret } from "cojson";
 
 export type BaseProfileShape = {
   name: z.core.$ZodString<string>;
@@ -84,7 +85,7 @@ export class AccountSchema<
       loadAs?: Account | AnonymousJazzAgent;
       resolve?: RefsToResolveStrict<AccountSchema<Shape>, R>;
     },
-  ): Promise<MaybeLoaded<Loaded<AccountSchema<Shape>, R>>> {
+  ): Promise<Settled<Loaded<AccountSchema<Shape>, R>>> {
     // @ts-expect-error
     return this.coValueClass.load(
       id,
@@ -103,8 +104,14 @@ export class AccountSchema<
         worker: Account,
       ) => Promise<void>;
     },
+  ): Promise<{
+    credentials: {
+      accountID: string;
+      accountSecret: AgentSecret;
+    };
     // @ts-expect-error we can't statically enforce the schema's resolve query is a valid resolve query, but in practice it is
-  ): Promise<Loaded<AccountSchema<Shape>, DefaultResolveQuery>> {
+    account: Loaded<AccountSchema<Shape>, DefaultResolveQuery>;
+  }> {
     // @ts-expect-error
     return this.coValueClass.createAs(worker, options);
   }