+++ examples/jazz-sveltekit/CHANGELOG.md
@@ -1,5 +1,43 @@
 # jazz-sveltekit
 
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies [68acca4]
+- Updated dependencies [c00a454]
+  - jazz-tools@0.19.11
+
+## 0.0.30
+
+### Patch Changes
+
+- jazz-tools@0.19.10
+
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies [d901caa]
+- Updated dependencies [a2bb9f0]
+  - jazz-tools@0.19.9
+
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies [21f7d34]
+- Updated dependencies [b22ad89]
+- Updated dependencies [28b23dd]
+  - jazz-tools@0.19.8
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [e113a79]
+  - jazz-tools@0.19.7
+
 ## 0.0.26
 
 ### Patch Changes