+++ packages/jazz-run/CHANGELOG.md
@@ -1,5 +1,64 @@
 # jazz-run
 
+## 0.19.11
+
+### Patch Changes
+
+- Updated dependencies [68acca4]
+- Updated dependencies [c00a454]
+  - jazz-tools@0.19.11
+  - cojson@0.19.11
+  - jazz-webhook@0.19.11
+  - cojson-storage-sqlite@0.19.11
+  - cojson-transport-ws@0.19.11
+
+## 0.19.10
+
+### Patch Changes
+
+- cojson@0.19.10
+- cojson-storage-sqlite@0.19.10
+- cojson-transport-ws@0.19.10
+- jazz-tools@0.19.10
+- jazz-webhook@0.19.10
+
+## 0.19.9
+
+### Patch Changes
+
+- Updated dependencies [d901caa]
+- Updated dependencies [a2bb9f0]
+  - jazz-tools@0.19.9
+  - cojson@0.19.9
+  - jazz-webhook@0.19.9
+  - cojson-storage-sqlite@0.19.9
+  - cojson-transport-ws@0.19.9
+
+## 0.19.8
+
+### Patch Changes
+
+- Updated dependencies [21f7d34]
+- Updated dependencies [b22ad89]
+- Updated dependencies [93e4a34]
+- Updated dependencies [28b23dd]
+  - jazz-tools@0.19.8
+  - cojson@0.19.8
+  - jazz-webhook@0.19.8
+  - cojson-storage-sqlite@0.19.8
+  - cojson-transport-ws@0.19.8
+
+## 0.19.7
+
+### Patch Changes
+
+- Updated dependencies [e113a79]
+  - jazz-tools@0.19.7
+  - jazz-webhook@0.19.7
+  - cojson@0.19.7
+  - cojson-storage-sqlite@0.19.7
+  - cojson-transport-ws@0.19.7
+
 ## 0.19.6
 
 ### Patch Changes