+++ packages/jazz-tools/CHANGELOG.md
@@ -1,5 +1,59 @@
 # jazz-tools
 
+## 0.19.11
+
+### Patch Changes
+
+- 68acca4: Sync Server is now configurable in inspector.jazz.tools
+- c00a454: Resolve footgun in Account.createAs() that made it easy to save the wrong credentials.
+- Updated dependencies [c00a454]
+  - cojson@0.19.11
+  - cojson-storage-indexeddb@0.19.11
+  - cojson-transport-ws@0.19.11
+
+## 0.19.10
+
+### Patch Changes
+
+- cojson@0.19.10
+- cojson-storage-indexeddb@0.19.10
+- cojson-transport-ws@0.19.10
+
+## 0.19.9
+
+### Patch Changes
+
+- d901caa: Added cojson-core-rn that improves ReactNative crypto performance
+- a2bb9f0: Added Suspense hooks for React and implemented subscription deduplication for React hooks
+- Updated dependencies [d901caa]
+  - cojson@0.19.9
+  - cojson-storage-indexeddb@0.19.9
+  - cojson-transport-ws@0.19.9
+
+## 0.19.8
+
+### Patch Changes
+
+- 21f7d34: Improve error logging in subscriptions and add stacktraces on errors coming from React hooks.
+
+  Added jazzConfig.setCustomErrorReporter API to intercept subscription errors and send them to an error tracker.
+
+- b22ad89: Narrow down `.load` return type to not include `loading` CoValues
+- 28b23dd: Added polyfills helper to React Native and Expo exports
+- Updated dependencies [93e4a34]
+  - cojson@0.19.8
+  - cojson-storage-indexeddb@0.19.8
+  - cojson-transport-ws@0.19.8
+
+## 0.19.7
+
+### Patch Changes
+
+- e113a79: Avoid migrating unauthorized CoValues
+  - cojson@0.19.7
+  - cojson-storage-indexeddb@0.19.7
+  - cojson-transport-ws@0.19.7
+
 ## 0.19.6
 
 ### Patch Changes
@@ -323,7 +377,6 @@
 ### Patch Changes
 
 - f2f478a: Add connection status API for React and Svelte
-
   - **React**: Added `useSyncConnectionStatus()` hook that returns the current connection status to the Jazz sync server
   - **Svelte**: Added `SyncConnectionStatus` class that provides reactive connection status monitoring
 
@@ -359,7 +412,6 @@
 ### Patch Changes
 
 - a584ab3: Add WasmCrypto support for Cloudflare Workers and edge runtimes by importing `jazz-tools/load-edge-wasm`.
-
   - Enable WasmCrypto functionality by initializing the WebAssembly environment with the import: `import "jazz-tools/load-edge-wasm"` in edge runtimes.
   - Guarantee compatibility across Cloudflare Workers and other edge runtime environments.
 
@@ -743,7 +795,6 @@
 - 3cd1586: Makes the key rotation not fail when child groups are unavailable or their readkey is not accessible.
 
   Also changes the Group.removeMember method to not return a Promise, because:
-
   - All the locally available child groups are rotated immediately
   - All the remote child groups are rotated in background, but since they are not locally available the user won't need the new key immediately